refactor: reverted back to custom-data module
The custom-data module now has two template scripts, one for Ubuntu and one for Windows. It's up to the user of GATR to determine which they want to use and ensure the correct script is specified. This will be documented in the README and made apparent when wanting to use Windows Runners.