feat(runner-controller): Add public IP address to network interface
This commit adds a new (temporary) feature to the `runner-controller` by including a public IP address to the network interface. The public IP address is created with a basic SKU and regional tier.

I'm adding this Public IP because I need access in order to debug an issue with the Windows Init script for GitHub Runners.