feat(runner-controller): Add public IP address to network interface creation
This commit adds a new feature to the runner-controller by including the creation of a public IP address when creating a network interface. The public IP address is created with a name and SKU, which are specified in the code.

This change is specifically so I'm able to access the VMs that are spun up by GATR. I need this access to debug why the `runner-init` script for Windows is not completing successfully.